requirements.txt
requests #==2.31.0
beautifulsoup4 #==4.12.2
lxml #==4.9.3
python-dotenv #==1.0.0
python-box #==7.1.1
pyyaml #==6.0.1
pandas #==2.1.1
supabase #==2.0.3
black #==23.9.1
jupyter #==1.0.0
tqdm #==4.66.1
colorama